Initial gdx-backend-defrac implementation
The defrac backend is non-functional in the sense that it misses
most of the functionality. Even the Graphics implementation does
not support al OpenGL calls.

The backend will stay this way for now since libgdx relies on
blocking IO and java.io.File methods which cannot be mapped to the
platforms supported by defrac, like HTML5.